requirements.txt
# Flask backend dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Audio processing
pyaudio==0.2.13
assemblyai==0.17.0

# AI model clients
mistralai==0.0.9
google-generativeai==0.3.1

# Utility packages
requests==2.31.0
numpy==1.24.4
pydub==0.25.1
